requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
alembic==1.13.2
passlib==1.7.4
bcrypt==4.2.0
email-validator==2.3.0
python-jose[cryptography]==3.3.0